backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
joblib==1.4.2
scikit-learn==1.5.2
Pillow==11.0.0
pytesseract==0.3.13
numpy==2.1.3
pandas==2.2.3
aiohttp==3.11.10
python-dotenv==1.0.1
email-validator==2.1.0

# MongoDB dependencies
motor==3.6.0
beanie==1.27.0
dnspython==2.7.0tpx
